+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/common/ToggleableFilterChip.kt
@@ -1,8 +1,13 @@
 package de.connect2x.messenger.compose.view.common
 
-import androidx.compose.material3.FilterChip
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedFilterChip
 import kotlinx.coroutines.flow.MutableStateFlow
 
 @Composable
@@ -13,8 +18,12 @@ fun <T> ToggleableFilterChip(
 ) {
     val applied = appliedFilters.collectAsState().value.containsAll(filters)
 
-    FilterChip(
+    ThemedFilterChip(
+        style = MaterialTheme.components.primaryChip,
         selected = applied,
+        leadingIcon = if (applied) {
+            @Composable { Icon(Icons.Default.Check, contentDescription = null) }
+        } else null,
         onClick = {
             if (applied) {
                 appliedFilters.value = appliedFilters.value.minus(filters)
@@ -22,8 +31,6 @@ fun <T> ToggleableFilterChip(
                 appliedFilters.value = appliedFilters.value.plus(filters)
             }
         },
-        label = @Composable {
-            label()
-        }
+        label = label
     )
 }