+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/settings/UserProfile.kt
@@ -16,7 +16,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
@@ -34,8 +33,6 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.SuggestionChip
-import androidx.compose.material3.SuggestionChipDefaults
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.minimumInteractiveComponentSize
@@ -77,6 +74,8 @@ import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.theme.components
 import de.connect2x.messenger.compose.view.theme.components.ThemedButton
 import de.connect2x.messenger.compose.view.theme.components.ThemedIconButton
+import de.connect2x.messenger.compose.view.theme.components.ThemedInfoChip
+import de.connect2x.messenger.compose.view.theme.components.ThemedSuggestionChip
 import de.connect2x.messenger.compose.view.theme.components.ThemedUserAvatar
 import de.connect2x.messenger.compose.view.theme.messengerDpConstants
 import de.connect2x.trixnity.messenger.viewmodel.room.settings.ChangePowerLevelViewModel
@@ -128,10 +127,7 @@ class UserProfileViewImpl : UserProfileView {
         val canOpenChat = userProfileViewModel.canOpenChat.collectAsState().value
 
         Column(
-            Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background)
-                .blockPointerInput(),
+            Modifier.fillMaxSize(),
             verticalArrangement = Arrangement.SpaceBetween
         ) {
             DialogHandler(userProfileViewModel)
@@ -151,9 +147,9 @@ class UserProfileViewImpl : UserProfileView {
                             Box(Modifier.align(Alignment.Center)) {
                                 ThemedUserAvatar(
                                     userInfoElement.initials,
-                                    image,
+                            image,
                                     this@BoxWithConstraints.maxWidth.coerceAtMost(200.dp)
-                                )
+                        )
                             }
                         }
                         Spacer(Modifier.height(20.dp))
@@ -168,17 +164,36 @@ class UserProfileViewImpl : UserProfileView {
 
                     Spacer(Modifier.height(5.dp))
                     when (userTrustLevel) {
-                        is UserTrustLevel.CrossSigned ->
-                            StatusRow(i18n.secure(), true) { VerifiedIcon(VerificationLevel.USER) }
+                        is UserTrustLevel.CrossSigned -> {
+                            ThemedInfoChip(
+                                style = MaterialTheme.components.primaryChip,
+                                icon = { VerifiedIcon(VerificationLevel.USER) },
+                                label = { Text(i18n.secure()) },
+                            )
+                        }
 
-                        is UserTrustLevel.NotAllDevicesCrossSigned ->
-                            StatusRow(i18n.insecure(), false) { NotVerifiedIcon(VerificationLevel.USER) }
+                        is UserTrustLevel.NotAllDevicesCrossSigned -> {
+                            ThemedInfoChip(
+                                style = MaterialTheme.components.destructiveChip,
+                                icon = { NotVerifiedIcon(VerificationLevel.USER) },
+                                label = { Text(i18n.insecure()) },
+                            )
+                        }
 
-                        UserTrustLevel.Blocked, is UserTrustLevel.Invalid, UserTrustLevel.Unknown, null ->
-                            StatusRow(i18n.roomNoEncryptionFound(), false) { NotVerifiedIcon(VerificationLevel.USER) }
+                        UserTrustLevel.Blocked, is UserTrustLevel.Invalid, UserTrustLevel.Unknown, null -> {
+                            ThemedInfoChip(
+                                style = MaterialTheme.components.destructiveChip,
+                                icon = { NotVerifiedIcon(VerificationLevel.USER) },
+                                label = { Text(i18n.roomNoEncryptionFound()) },
+                            )
+                        }
                     }
                     if (membership == Membership.BAN) {
-                        StatusRow(membershipReason ?: i18n.banned(), false) { BanIcon() }
+                        ThemedInfoChip(
+                            style = MaterialTheme.components.destructiveChip,
+                            icon = { BanIcon() },
+                            label = { Text(membershipReason ?: i18n.banned()) },
+                        )
                     }
                 }
             }
@@ -573,7 +588,7 @@ fun ChangingPowerLevel(userProfileViewModel: UserProfileViewModel) {
                 )
                 FlowRow(Modifier.fillMaxWidth(), Arrangement.SpaceBetween, Arrangement.Top) {
                     if (canSetRoleToUser) {
-                        SuggestionChip(
+                        ThemedSuggestionChip(
                             onClick = {
                                 changePowerLevelInput = TextFieldValue(
                                     ChangePowerLevelViewModel.Role.USER
@@ -586,7 +601,7 @@ fun ChangingPowerLevel(userProfileViewModel: UserProfileViewModel) {
                         )
                     }
                     if (canSetRoleToModerator) {
-                        SuggestionChip(
+                        ThemedSuggestionChip(
                             onClick = {
                                 changePowerLevelInput = TextFieldValue(
                                     ChangePowerLevelViewModel.Role.MODERATOR
@@ -599,7 +614,7 @@ fun ChangingPowerLevel(userProfileViewModel: UserProfileViewModel) {
                         )
                     }
                     if (canSetRoleToAdmin) {
-                        SuggestionChip(
+                        ThemedSuggestionChip(
                             onClick = {
                                 changePowerLevelInput = TextFieldValue(
                                     ChangePowerLevelViewModel.Role.ADMIN
@@ -671,34 +686,6 @@ private fun MenuElement(
     ) {
         content()
     }
-}
-
-@Composable
-private fun StatusRow(text: String, positive: Boolean = true, icon: @Composable () -> Unit) {
-    SuggestionChip(
-        enabled = false,
-        onClick = {},
-        label = {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(MaterialTheme.messengerDpConstants.verySmall),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                icon()
-
-                BoxWithConstraints {
-                    Text(
-                        text,
-                        color = if (positive) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onErrorContainer,
-                        modifier = Modifier.widthIn(max = maxWidth * .5f)
-                    )
-                }
-            }
-        },
-        colors = SuggestionChipDefaults.suggestionChipColors().copy(
-            disabledContainerColor = if (positive) MaterialTheme.colorScheme.primaryContainer else MaterialTheme.colorScheme.errorContainer,
-            disabledLabelColor = if (positive) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onErrorContainer
-        ),
-    )
 }
 
 @Composable